test: split RTSP seek-mode coverage into its own file
test_rtsp_misc.py was approaching 1300 lines, ~70% of which covered
r2h-seek-mode opt-in semantics, recent-clock playback, and the
M3U-configured query-merge precedence rule.

Move TestRTSPRecentPlayseek, TestRTSPSeekMode, and
TestRTSPSeekModeQueryMerge — together with the helpers they own
(_format_basic_utc, _format_yyyyMMddHHmmss, _expected_clock_str) —
into test_rtsp_seek_mode.py with its own module-scoped shared_r2h
fixture.

No test logic changes; 51 tests still collect and pass.
